Move near-events to omni-types
